% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/film_colors.R
\name{film_colors}
\alias{film_colors}
\title{A Film Color Palette Generator}
\usage{
film_colors(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{Vertigo}}

\item{n}{Number of colors desired. Unfortunately most palettes now only
have 4 or 5 colors. If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.}
}
\value{
A vector of colours.
}
\description{
These are a handful of color palettes from movies.
}
\examples{
film_colors("Vertigo")

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
pal <- film_colors(21, name = "Vertigo", type = "continuous")
image(volcano, col = pal)
}
\keyword{colors}

