% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_colors.R
\docType{data}
\name{film_palettes}
\alias{film_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 1.}
\usage{
film_palettes
}
\description{
Use \code{\link{film_colors}} to construct palettes of desired length.
}
\keyword{datasets}

