% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_colors.R
\name{sort_colors}
\alias{sort_colors}
\title{A Color Sorting Function}
\usage{
sort_colors(col)
}
\arguments{
\item{col}{vector of any of the three kinds of \R color specifications, i.e., either a color name (as listed by \code{\link{colors}()}), a hexadecimal string of the form \code{"#rrggbb"} or \code{"#rrggbbaa"} (see \code{\link{rgb}}), or a positive integer \code{i} meaning \code{\link{palette}()[i]}.}
}
\value{
A vector of colors.
}
\description{
A Color Sorting Function
}
\examples{
colors <- c("#060306", "#290705", "#131325")
sorted_colors <- sort_colors(colors)
sorted_colors
}
\references{
\url{https://learnr.wordpress.com/2010/01/21/ggplot2-crayola-crayon-colours/}
}

